DP-214 - Update roles within the Stub API to be lower case - Moved to Serialisation folder
